shared_mem: rename histogram → shared_mem (pattern dir)
git mv examples/histogram → examples/shared_mem (example_rename.md Phase 3): the
directory names the shared-memory pattern; the histogram computation keeps its
identity in the files inside (hist.cpp, hist_demo.py, HistAccel/HistTest, the
pysilicon_hist_proj Vitis project — all unchanged).

Update every live reference to the example-directory path: the test imports
(examples.histogram → examples.shared_mem) and HIST_EXAMPLE_DIR path component,
the root .gitignore artifact patterns, the docs/guide links (memory, timing),
and cosmetic comments in hwgen.py / increment notes / mcp resources. Add a
docs/examples/shared_mem/ index page describing the pattern and slot it into the
examples nav at teaching position 4; the example index now links it.

mcp/corpus/ and the computation-named project dir are intentionally untouched.
