+++ examples/shared_mem/hist.cpp
