+++ examples/shared_mem/hist.hpp
