+++ examples/shared_mem/hist_demo.py
