+++ examples/shared_mem/hist_tb.cpp
