+++ examples/shared_mem/run.tcl
